+++ luncher_backend/users/admin.py
@@ -1,5 +1,7 @@
 from django.contrib import admin
+from .models import Location
 from .models import User
 
 
 admin.site.register(User)
+admin.site.register(Location)
